+++ src/http/index.ts
@@ -3,6 +3,7 @@ export type {
   HttpRequestConfig,
   HttpResponse,
 } from './client.ts'
+export type { HttpErrorRequestConfig } from './errors.ts'
 
 export { HttpClient, readHeaders } from './client.ts'
 export { HttpError, isHttpError } from './errors.ts'